Extract processor to its own file.
